Rewrite of PrimaryKey component - Rewritten and unified PrimaryKey component - Moved bookmark button inside PrimaryKey
