--- src/app/components/modal/components/read/PrimaryKey.component.html
@@ -1,48 +1,0 @@
-<ion-card>
-    <div
-        style="width: 50px; height: 100%; position: absolute"
-        [style.background-color]="
-            tagConfig && tagConfig.markerColor ? tagConfig.markerColor : 'black'
-        "
-    >
-        <app-icon
-            class="iconImgPNG imgCenter"
-            [icon]="tagConfig && tagConfig.icon ? tagConfig.icon : null"
-            [jsonSprites]="jsonSprites"
-        >
-        </app-icon>
-    </div>
-    <div class="primaryCardContent">
-        <ion-card-content>
-            <p class="primaryKeyLabel" *ngIf="!displayCode && tagConfig?.lbl">
-                {{
-                    tagConfig.lbl[language]
-                        ? tagConfig.lbl[language]
-                        : tagConfig.lbl['en']
-                }}
-            </p>
-
-            <p
-                class="description"
-                *ngIf="!displayCode && tagConfig && tagConfig.description"
-            >
-                {{
-                    tagConfig.description[language] !== undefined
-                        ? tagConfig.description[language]
-                        : tagConfig.description['en']
-                }}
-            </p>
-
-            <span *ngIf="displayCode && tagConfig">
-                <ion-icon name="code"></ion-icon>
-                {{ tagConfig.tags | displayTags }}
-                <!-- {{this.tagsConfig[primaryKey.key].lbl[language] || this.tagsConfig[primaryKey.key].lbl['en'] || ''}} -->
-            </span>
-
-            <span *ngIf="!tagConfig">
-                <ion-icon name="code"></ion-icon>
-                {{ primaryKey.key }}={{ primaryKey.value }}
-            </span>
-        </ion-card-content>
-    </div>
-</ion-card>
